+++ libs/base_core/src/text/char/mod.rs
@@ -8,7 +8,6 @@ mod chars; // CharIter
 mod definitions; // char7, char8, char16, char_utf8
 mod namespace; // Char
 mod reexports;
-mod unicode_scalar; // UnicodeScalar
 
 // without re-exports
 mod impls;
@@ -23,7 +22,6 @@ crate::structural_mods! { // _mods
             definitions::*,
             namespace::*,
             reexports::*,
-            unicode_scalar::*,
         };
     }
 }